refactor: split model from field_list_controller.py
